+++ changelog.md
@@ -1,2 +1,2 @@
 ### Benched 1.2.2
-- Initial release of Benched for NeoForge
+- Initial release of Benched for Minecraft 1.20.5 & 1.20.6